Moved Indexer to formats.sql
