+++ frictionless/formats/sql/__init__.py
@@ -1,5 +1,6 @@
 from .adapter import SqlAdapter
 from .control import SqlControl
+from .indexer import SqlIndexer
 from .mapper import SqlMapper
 from .parser import SqlParser
 from .plugin import SqlPlugin
